+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 
+### Added
+
+- Add support of single file in addition to folder in SasFileProvider [#265](https://github.com/nbbrd/jdemetra-sas/issues/265)
+
 ## [2.2.2] - 2024-09-03
 
 This is a bugfix release of **jdemetra-sas**.